+++ config.h
@@ -196,6 +196,7 @@ struct config {
         uint64_t delayed_render_lower_ns;
         uint64_t delayed_render_upper_ns;
         off_t max_shm_pool_size;
+        float box_drawing_base_thickness;
     } tweak;
 
     user_notifications_t notifications;